requirements.txt
Django
djangorestframework
psycopg2-binary
pytest
pytest-django
model_bakery
pytest-cov